DBZ-8825 Move methods for adding fields into ConnectRecordUtil class
There's not changes so far, just moving the methods, adjusting their
parameters and made them static.
